event-helper: Create a new system event helper library
The system event helper library provides system event management, so
that the tasks can wait on the required system event to happen.
